+++ scripts/release/debian/build.sh
@@ -35,20 +35,8 @@ make install
 
 export PATH=$PATH:$WORKDIR/python_env/bin
 
-# note: This installation step could happen in debian/rules but was unable to escape $ char.
-# It does not affect the built .deb file though.
-pip3 install wheel
-
-if [ -d $WORKDIR/privates ]; then
-    # The private pages are used when a dependency is not published yet
-    find $WORKDIR/private -name '*.whl' | xargs pip3 install
-fi
-
-# The product whl are expected to be pre-built
-find /mnt/artifacts/pypi -name '*.whl' | xargs pip3 install
-
-pip3 install --force-reinstall --upgrade azure-nspkg azure-mgmt-nspkg
-pip3 install --force-reinstall urllib3==1.24.2
+find ${WORKDIR}/src/ -name setup.py -type f | xargs -I {} dirname {} | grep -v azure-cli-testsdk | xargs pip3 install --no-deps
+pip3 install -r ${WORKDIR}/src/azure-cli/requirements.py3.$(uname).txt
 
 # Create create directory for debian build
 mkdir -p $WORKDIR/debian