+++ src/azure-cli/azure/cli/command_modules/network/_params.py
@@ -479,6 +479,7 @@ def load_arguments(self, _):
     with self.argument_context('network dns record-set ns') as c:
         c.argument('dname', options_list=['--nsdname', '-d'], help='Name server domain name.')
         c.argument('subscription_id', options_list=['--subscriptionid', '-s'], help='Subscription id to add name server record')
+        c.ignore('_subscription')
 
     with self.argument_context('network dns record-set ptr') as c:
         c.argument('dname', options_list=['--ptrdname', '-d'], help='PTR target domain name.')