+++ src/azure-cli/azure/cli/command_modules/vm/_validators.py
@@ -1090,11 +1090,6 @@ def _get_default_address_pool(cli_ctx, resource_group, balancer_name, balancer_t
     return values[0]
 
 
-def _validate_vmss_zone_args(namespace):
-    if namespace.platform_fault_domain_count is not None and namespace.zones is None:
-        raise CLIError('usage error: --platform-fault-domain-count COUNT --zones ZONES')
-
-
 def _validate_vmss_create_load_balancer_or_app_gateway(cmd, namespace):
     from msrestazure.azure_exceptions import CloudError
     from msrestazure.tools import parse_resource_id
@@ -1233,7 +1228,6 @@ def process_vmss_create_namespace(cmd, namespace):
     _validate_vm_create_storage_profile(cmd, namespace, for_scale_set=True)
     _validate_vm_vmss_create_vnet(cmd, namespace, for_scale_set=True)
 
-    _validate_vmss_zone_args(namespace)
     _validate_vmss_create_load_balancer_or_app_gateway(cmd, namespace)
     _validate_vmss_create_subnet(namespace)
     _validate_vmss_create_public_ip(cmd, namespace)